+++ test/end-to-end/signin-rrdv6.test.js
@@ -51,6 +51,7 @@ import {
     waitForSTElement,
     backendBeforeEach,
     getInvalidClaimsJSON,
+    expectErrorThrown,
 } from "../helpers";
 import fetch from "isomorphic-fetch";
 import { SOMETHING_WENT_WRONG_ERROR } from "../constants";
@@ -608,7 +609,7 @@ describe("SuperTokens SignIn with react router dom v6", function () {
             ]);
         });
 
-        it("Should redirect to onFailureRedirections result if it's a relative path", async function () {
+        it("Should throw if onFailureRedirections returns a relative path", async function () {
             await Promise.all([
                 page.goto(
                     `${TEST_CLIENT_BASE_URL}/auth?redirectToPath=${encodeURIComponent("/dashboard?test=value#asdf")}`
@@ -626,16 +627,13 @@ describe("SuperTokens SignIn with react router dom v6", function () {
                 page.waitForNavigation({ waitUntil: "networkidle0" }),
             ]);
 
-            await page.evaluate(() => {
-                const validator = window.UserRoleClaim.validators.includes("admin");
-                validator.onFailureRedirection = () => "second-factor";
-                window.setClaimValidators([validator]);
+            await expectErrorThrown(page, async () => {
+                await page.evaluate(() => {
+                    const validator = window.UserRoleClaim.validators.includes("admin");
+                    validator.onFailureRedirection = () => "second-factor";
+                    window.setClaimValidators([validator]);
+                });
             });
-
-            await page.waitForNavigation({ waitUntil: "networkidle0" });
-
-            let href = await page.evaluate(() => window.location.pathname);
-            assert.strictEqual(href, "/dashboard/second-factor");
         });
 
         it("Should redirect to onFailureRedirections result if it's on another domain", async function () {